rename template pallet to livecoin
