+++ pallets/livecoin/src/lib.rs
@@ -1,0 +1,84 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+
+/// A FRAME pallet template with necessary imports
+
+/// Feel free to remove or edit this file as needed.
+/// If you change the name of this file, make sure to update its references in runtime/src/lib.rs
+/// If you remove this file, you can remove those references
+
+/// For more guidance on Substrate FRAME, see the example pallet
+/// https://github.com/paritytech/substrate/blob/master/frame/example/src/lib.rs
+use frame_support::{decl_error, decl_event, decl_module, decl_storage, dispatch};
+use system::ensure_signed;
+
+#[cfg(test)]
+mod mock;
+
+#[cfg(test)]
+mod tests;
+
+/// The pallet's configuration trait.
+pub trait Trait: system::Trait {
+    // Add other types and constants required to configure this pallet.
+
+    /// The overarching event type.
+    type Event: From<Event<Self>> + Into<<Self as system::Trait>::Event>;
+}
+
+// This pallet's storage items.
+decl_storage! {
+    trait Store for Module<T: Trait> as Livecoin {
+        TotalSupply get(fn total_supply): u64;
+    }
+}
+
+// The pallet's events
+decl_event!(
+    pub enum Event<T>
+    where
+        AccountId = <T as system::Trait>::AccountId,
+    {
+        SupplyChanged(u64, AccountId),
+    }
+);
+
+// The pallet's errors
+decl_error! {
+    pub enum Error for Module<T: Trait> {
+        /// Value was None
+        NoneValue,
+        /// Value reached maximum and cannot be incremented further
+        StorageOverflow,
+    }
+}
+
+// The pallet's dispatchable functions.
+decl_module! {
+    /// The module declaration.
+    pub struct Module<T: Trait> for enum Call where origin: T::Origin {
+        // Initializing errors
+        // this includes information about your errors in the node's metadata.
+        // it is needed only if you are using errors in your pallet
+        type Error = Error<T>;
+
+        // Initializing events
+        // this is needed only if you are using events in your pallet
+        fn deposit_event() = default;
+
+        /// Just a dummy entry point.
+        /// function that can be called by the external world as an extrinsics call
+        /// takes a parameter of the type `AccountId`, stores it, and emits an event
+        pub fn set_supply(origin, new_supply: u64) -> dispatch::DispatchResult {
+            // Check it was signed and get the signer. See also: ensure_root and ensure_none
+            let who = ensure_signed(origin)?;
+
+            // Code to execute when something calls this.
+            // For example: the following line stores the passed in u32 in the storage
+            TotalSupply::put(new_supply);
+
+            // Here we are raising the Something event
+            Self::deposit_event(RawEvent::SupplyChanged(new_supply, who));
+            Ok(())
+        }
+    }
+}
