+++ runtime/src/lib.rs
@@ -37,8 +37,8 @@ pub use frame_support::{
 	weights::Weight,
 };
 
-/// Importing a template pallet
-pub use template;
+/// Importing the livecoin pallet
+pub use livecoin;
 
 /// An index to a block.
 pub type BlockNumber = u32;
@@ -221,8 +221,7 @@ impl sudo::Trait for Runtime {
 	type Call = Call;
 }
 
-/// Used for the module template in `./template.rs`
-impl template::Trait for Runtime {
+impl livecoin::Trait for Runtime {
 	type Event = Event;
 }
 
@@ -240,8 +239,7 @@ construct_runtime!(
 		Balances: balances::{Module, Call, Storage, Config<T>, Event<T>},
 		TransactionPayment: transaction_payment::{Module, Storage},
 		Sudo: sudo::{Module, Call, Config<T>, Storage, Event<T>},
-		// Used for the module template in `./template.rs`
-		TemplateModule: template::{Module, Call, Storage, Event<T>},
+		Livecoin: livecoin::{Module, Call, Storage, Event<T>},
 	}
 );
